Added functionality for tiptap and styling for markdown

+++ src/app/reviews/post-review/actions.ts
@@ -22,7 +22,7 @@ export async function postReview(formData: FormData) {
             category,
             description: description.trim(),
             location,
-            ratingAvg: parseInt(ratingAvg),
+            ratingAvg: parseFloat(ratingAvg),
             title: title.trim(),
             verified: Boolean(verified),
             reviewImageUrl: uploadedReviewImageUrl,
